feat(e2e): add triage analysis and session management tests
- Implemented end-to-end tests for triage analysis with rule-based fallback.
- Added tests for managing triage sessions, including creation, cancellation, and progress tracking.
- Removed outdated triage MG tests to streamline the test suite.
